Update the .gitignore file
Because I use IDEA and Visual Studio Code, I don't want to pollute
the repository with useless configuration files.

 - IDEA saves all its files into the .idea directory
 - VSCode saves all its files into the .vscode directory

These two directories have to be ignored by git.

Finally, IDEA automatically compile the source code (via CMake)
in a cmake-build-debug or cmake-build-release directory. To never
accidentally push binaries and build files, these directories are
ignored. The same applies with the build directory when the
compilation is done in cli.