Add CMake toolchain and Logger class implementation
Ektor is divided in small static libraries, one library for each
module. This commit add all CMakeFiles to be able to compile
these modules.

The first module, Utils, is also added. It only contains the class
Logger that allows to centralise all log messages. For now, these
messages are dumped on the standard output (and error output), but
could be upgraded in the near future to populate a external log
file with journalisation and rotation.

This commit doesn't allow the library to compile successfully.
The Memory module is still missing because it needs a little more
local testing.