Add StackAllocator implementation
The StackAllocator implementation is a linked list of memory pages
containing a stack of allocated slots. During a new allocation,
the top slot (free) is returned. Pages can be expanded by
allocating a new one. During deallocation, slots are unstacked and
if a page is completely unused, it's destroyed.

There is always at least one page available to avoid an obvious
malloc on the first alloc.

Warning: this code is not thread-safe and probably never will!

This commit doesn't fix the build problem. I'm cleaning the demo
code locally.