Add FrameAllocator implementation
The FrameAllocator implemention is a linked list of fixed size
memory frames. During a new allocation, a new frame is returned.
On a desallocation, the frame is pushed back in the linked list.

All pages (containers of frames) are freed during the allocator
destruction, so the developer has to take in account the large
FrameAllocator memory footprint.

Because this allocator only manage fixed size frames, if an
allocation larger than a frame is done, an exception is raised.

Warning: this code is not thread-safe and probably never will!

This commit breaks the library compilation. I'm working on a local
demo.