Add a FrameAllocator demo
This commit brings an example of FrameAllocator usage. It is based
on the StackAllocator demo and is very similar.

The FrameAllocator is perfect to create pools of objects or
factories.

However, their usage is very limited (cannot exceed the frame
size).

This commit also add an error when the limit is exceeded.