Add the Gfx module and Window implementation on Unix system
The Gfx module contains an implementation of Window that can be
used to open X11 window with an OpenGL 3.3 context.

To be compiled, it is necessary to have X11 installed on the
system.

Based on an OpenGL extension, not every computer can create such
context. If it's your case, please check your graphic card driver.

Simple events also have been implemented.